Fix some i18n errors in combination with multiple batteries
Also, default to show only one aggregate battery, as that's the
information most people need. And the remaining time for those
with more than one battery plugged in is fixed as well with this
commit.

Sorry translators for being such a pain :/ but even untranslated
strings are better than big, fat error messages in the UI


CCMAIL:kde-i18n-doc@kde.org

svn path=/branches/KDE/4.4/kdebase/workspace/; revision=1082224