backport r1073143:
* make the reading of disk usage actually async (e.g. we return between each device check)
* don't check every 10s (wtf?!), but only when requested; this moves it into the hands of the visualization

what this means is that plasma sleeps more, disks are accessed less, far fewer data members in the SystemModel class and far more readable code. probably too late for 4.4.0, which is unfortunate but my fault since it fell off my "to backport" list before it was actually backported. if there is movement in the tag in kdebase for other reasons, it would be nice if this was captured as well. distributions may also want to grab it for their 4.4 packages until 4.4.1 is out.

CCBUG:200184
CCMAIL:release-team@kde.org

svn path=/branches/KDE/4.4/kdebase/workspace/; revision=1087230