Fix Calendar Data Engine to return recurring event incidences, limit returned events to just the requested date range, and to dynamically update correctly.
The original implementation copied classes from kdepim/akonadi/kcal, this
change:
1) Uses some different classes from the same location that gives us the info we 
   need
2) Uses the latest verion of the copied code
3) Places the code in a sub-directory to make it obvious
4) Adds a README.txt explaining the code

These classes will be moved from kdepim to kdepimlibs in SC 4.6, after which 
time they will be used directly and this copy deleted.  These originals will
need to be monitored in the meantime to ensure no patches are lost.

Patches to the calendar widget to use the new recurrence dates will follow in a
a separate patch.


svn path=/branches/KDE/4.5/kdebase/workspace/; revision=1147188