Unbreak command history in krunner
The signal QLineEdit::textEdited misses changes involving the user selecting a new entry in the 
combobox or rotating history items with the appropriate shortcut keys.
This fixes (hopefully) all such problems by 
1) connecting to QComboBox::currentIndexChanged
2) providing an ad-hoc signal in the eventFilter triggered by the history rotation keys

This should not be forwardported to trunk; I'll implement a new widget for the history combobox
which hopefully will cleanly solve all such issues (and those related to switching focus to the 
scene using the arrow keys in some situations)

@dfaure:
Is it intended that KHistoryComboBox does not emit currentIndexChanged when 
rotating items with the shortcut keys?
CCMAIL:faure@kde.org


svn path=/branches/KDE/4.5/kdebase/workspace/; revision=1158323