backport of r1166775
Support X-KDE-System-Settings-Parent-Category-V2 entries also for categories in
a 3rd party usecase. KPackageKit for example distributes its own category and
wants to remain compatible with KDE <= 4.4.

What is note worthy that the if condition will be true if both the parent AND
the V2 entry are empty which is a) bogus and b) pointless since it would then
be a major entry (aka section) and have no need for a V2 ;)

svn path=/branches/KDE/4.5/kdebase/workspace/; revision=1166776