Fix window movment regression (e.g. pack) in 4.5.
The regression was caused in tiling which set the geometry to a different one again.
Nikhil please verify that tiling is still working correctly. Looking at the code this
commit should be side-effect free. When coming from Client::move() moveResizeGeom and
initialMoveResizeGeom may not be set. So I think the code in notifyTilingWindowMove()
cannot work at all ;-)
CCMAIL: nsm.nikhil@gmail.com
BUG: 241049
FIXED-IN: 4.5.2


svn path=/branches/KDE/4.5/kdebase/workspace/; revision=1169432