backporting r1169287
Make weather dataengine work if started without network

Always call setData() using the original source so it's initialized
and when we reset the ions, they have sources connected to them and are
able to communicate back.

Also, fix the ions to update their data into the weather engine and later
emit the signal that will make the engine force the clients (plasmoids) to
update using the new data.

This mostly fix the problem of turning the computer on without network,
getting network later and the weather plasmoids not updating.

CCBUG: 226022
CCBUG: 245639


svn path=/branches/KDE/4.5/kdebase/workspace/; revision=1170055