Backport r1190779.
Properly quote the call to 'export -p'.

Even though bash (and dash) do not differ 'export' and 'export -p', that call
was clearly wrong as the -p was being passed as a parameter to /bin/sh, not
export.

Thanks to M. Warner Losh for bringing that up.

CCMAIL: imp@ixsystems.com
CCMAIL: kde-freebsd@kde.org

svn path=/branches/KDE/4.5/kdebase/workspace/; revision=1190780