Plasma Calendar: Re-enable akonadi event updates
Fix the plasma calendar to set the right flag to indicate a refresh of
the akonadi events is required.  This should fix the problem, but while
I have confirmed this does now send a query to akonadi again, I don't
get any events back, but this si probably an artifact of my test
environment, it should work ok once in a full desktop/akonadi
environment.

CCBUG: 258125