adjustSize of the Stack Widget once the StackDialog is hidden
Right now a notification with a long title will create a very long
NotificationWidget and that widget will resize the stackdialog to
its size. The problem is that once the large notification group is
closed, the StackDialog will continue having the notification width
and it won't resize to a normal size.

This patch calls to "adjustSize" when the stackdialog doesn't
have any notification, so it will be resized to the right size.