revert wallpaper to "air"
why is this called oxygen-air and the screenshot shows an "air"
background? right: because the wallpaper is "ethais".
and why is this in kdebase? right again: because the wallpaper is in
kdeartwork.

uhm, maybe not right, after all. whatever the artistic vision behind
this is, the current state is technically bogus. consider moving this
theme to kdeartwork as a whole and making adjustments there.

BUG: 262554
FIXED-IN: 4.6.2