Don't segfault in release mode.
No idea why it's 0, but we're going to delete this copy of calendar.cpp soon,
so it's a waste of time to debugging it.

PIM applications don't crash at all regarding this, in plasma the ETM is probably
set incorrectly and has items with unsupported payloads.

BUG: 261593
(cherry picked from commit b2a322448e644cdf55b9990c90920ce8c0ffcc7f)