prevent crash when hiding the year boxes more than once
BUG:272077

QSpinBox emits the finished signal when enter is pressed and when
it loses focus. so if you hide it on enter, it then loses focus
and the signal is emitted again. even if the value hasn't changed
and no editing was done (not even returned from the previous signal
emission). moderately lame.