clear background pixmap before painting
this avoids that incomplete themes leave garbage on the screen.
this is an ugly workaround, but specifying a background for non-primary
screens is a somewhat new feature, and people typically wouldn't think
of it, so many themes are broken in this regard.
(cherry picked from commit 5ab6a7917852466be48958d3cccf128c9cceb1e5)

CCBUG: 268993