Initial multiple desktop support for Wayland clients
Through the window decoration it becomes possible to change the
desktop of the Wayland client. Desktop is very simply implemented
with an int property representing the desktop the client is on.
When desktop changes the decoration is shown/hidden accordingly.

In opposite to the X11 implementation neither the client nor
Workspace is informed about the change. We will have to see what
we need from that.