Wayland Clients can be activated
When setting to active the client gets the Wayland keyboard focus
and the decoration the X keyboard focus.
Currently the Client gets automatically focus when created - this
has to be moved to workspace somehow. Also the client does not lose
focus yet, so it stays in the active state even if another client
(Wayland or X) is the new active one.