In Wayland the window texture size does not include deco
In X11 the actual window is reparented into the decoration frame
causing the overal window pixmap to have the size of the window
plus the decoration.
This is not the case in Wayland where the decoration is completly
different surface. We need to keep this in mind when generating
the window quads and finally the texcoords from them.

Unfortunately this involves code removed in master.