Make blur work for windows with Force Blur set in fullscreen effects
Fixes a regression with the dashboard effect: the blur behind
setting was lost. Dashboard is a fullscreen effect, but does
not do full repaints, so we need to add the "normal" blur
handling for fullscreen effects if there is a window which
enforces blur behind even for fullscreen effects.

Remark: only needed in 4.7 branch, regression already fixed in other
way in master.

REVIEW: 101907