Don't load output config in Screen ctor since it can be wrong.
When the KCM is started, it will load the last used configuration, that
configuration can be wrong (have an old resolution, setup...), in that
case the KCM will load it and break your current setup.

This patch skip the loading of that part of the config making the KCM
NO intrusive, and behaving like it should meaning like a standar KCM
doing stuff only when Apply button is clicked.

Technically:
This is yet another hack, adding a parameter just to fullfill one
situation, but that's the best way I could imagine to fix the issue and
avoid to modify lot of code.
(cherry picked from commit f59d8bf4c4250ba54606d4420eeadd23653bc999)