Save proposed size into a config if the output is not connected instead
if we click on a size and the output is not active, KRandRTray will fail
to apply the setting and instead it will save it in a config file. This
is incorrect, what KRandR should be checking is if the output is
connected since an innactive but connected monitor can be configured.

Basically this commit fixes than when a resolution is click and the
output is innactive, it will work anyway.
(cherry picked from commit 2209720ebfd4c1b4a9c5e0bb6213ff08555f2d0b)