fix shouldShow to take into consideration actual GroupManager settings
previously, it was assuming that if the item wasn't on your desktop or activity, it shouldn't show.
obviously, the tasks widget can (and by default _is_) set up differently. this would cause launchers
to always show under such circumstances depending on which desktop the apps (or the desktop) were
started up on (e.g. at log in, or on plasma-desktop restart)