Adding an explicit ScreenLocker class
This class handles just the screen locking, but not the screen
saving. This are two different concepts which have to be separated.

The ScreenLocker takes care of locking the screen by e.g. the
shortcut and invokes for example the legacy screen saver engine to
show the saver and "lock" the screen.

The ScreenSaverEngine is hardly changed. It just forwards the lock
request to the ScreenLocker to let the locker decide how to lock.
The Saving functionality itself is not changed at all including the
D-Bus interface which has to stay in the saver as it is modeled
around screen saving and not screen locking. We probably need a new
fdo D-Bus interface for that.

In the next step it is possible to move the locking into an effect
to no longer require an X screen saver to perform the locking which
would immediately remove all the glitches we see with the screen
savers.

Also in a next step we need to ensure that the screen stays locked
even if KWin crashes (e.g. through a lock file).

Conflicts:

	kwin/CMakeLists.txt
	kwin/useractions.cpp
	kwin/workspace.cpp
	kwin/workspace.h