Kill Runner now does not use kdesu for killing processes as another user.
Instead of using kdesu for killing processes the kill runner now tries
to kill the process using kill with the current user's rights. In case
this does not work the KAuth::Action KSysguard uses to send signals to
processes as root is used to send SIGKILL resp. SIGTERM to the process.

In order to not have to lookup the signal ids in a map the signals are
now handled as ints and not as strings as done before.

(cherry picked from commit 020833b106fed0b1e19b9e01696006542461a4c6)