Change SystemModel to a single-hierarchy model
It used to be a tree view but badly implemented. Items from the proxied
model were present in all tree levels and filtered in data access by not
returning values for items not matching the corrent section. This is
rather bad for the usage in QML as tree views are hardly supported and
the row count is always incorrect.

This changes the model to no longer being a tree view. The mapping to
the proxied model is now correctly applied in mapFromSource and there
is no filtering any more performed in data.

The previous root elements became an additional role of the entries.

The section for fixed devices is dropped as it was unused. As well
it seems that the usage info is unused, therefore the code in data
is dropped. It might be an idea to also drop the remaining methods.