found more dead code
set the default theme when the combobox is changed. i wanted a way to
change it *just* for this shell, but i have a feeling it's more invasive
than that. i couldn't think of any methods to hack around it, either
(either through reimplementing corona, view, or containment..)

here's hoping my intuition is wrong.

i wish there were a way to say that the theme is being loaded, since it
takes a good few seconds, but it does it async it seems with no signals
in either direction