screw it, i'm doing the ui in code now, it's easier
the qt-designer is being a pain, and it's just far easier to do what i
want to do in code. going to make a tab widget on the right pane, which
will toggle b/w svg preview and plasma shell embedded.

basic layout functionality (only left pane is "done"), so the rest is
broken for now.