add a panel applet
silly me was wondering why my panel theme change wasn't working. turns
out i was looking at the task manager, which looks very similar to a
panel with a task manager inside it. panel is just blank, and it isn't a
containment unfortunately. i wasn't sure on how to add a containment in
a containment. they should just be widgets, so it should be easy. will
have to see later when i find time