ok, this is pretty cool...
delete the theme cache (*.kcache) everytime a different theme is
selected. in other words, you don't have to restart *anything* to be
able to see the changes in a theme. however, you still have to select a
different theme...going to be doing a KDirWatch fix for that, so it'll
happen autofreakingmagically