attempt to fix painting.
this is badly in need of a refactoring. when the plotter begins shifting
left, hovering is totally gone.this is because painted is triggered on
an interval, when in reality it should be when needed as well. i'm not
sure how to break this up properly, i feel like a separate canvas just
for the text is overkill. but it needs access to the points[]