Add Tilt-Wheel, "XButton1", "XButton2" support for Kickoff
Workaround removal of the "Back Button" (in 4.7): In the
"Applications" GUI, users with tilt-wheel mice can use
"Tilt-Left" as a "Back Button" equivalent. If the focused
item is a Parent for other items, "Tilt-Right" will move
to the right, expanding the Children.

For "gamer" mice with thumb buttons, "XButton1" (AKA the
"back" button) will work just like the "Tilt-Left" Wheel
Event; and "XButton2" will work just like "Tilt-Right".

BUG: 289519
REVIEW: 104406