Unbreak Python 2 applet loading.
exec() is a function in Py3, but the syntax used in this file does not
allow it for being called as a function in Python 2. Thus, a simple check
is made, calling the right function depending on the Python version.

Anyone who has Python3 installed, please confirm that it works.

CCMAIL: martin.sandsmark@kde.org