Improve Drop behaviour
When a favorite is dropped under the list, the drop doesn't succeed. The
reason is that the drop is only caught by the items, so there's nothing
catching it when it goes somewhere where there's no item.

Reproduce: in Favorites, drag an item to the bottom of the list. It
works as long as you drop it onto the last item, but silently fails
if/when you drop it under the last item.

We work around this by creating a variable size, invisible footer item
which catches the drop.

This patch also makes the list non-flickable if it doesn't have to,
and fixes a crash when the new list index of the item is bigger than the
number of items.