Unbreak Python 2 applet loading.
exec() is a function in Py3, but the syntax used in this file does not
allow it for being called as a function in Python 2. Thus, a simple check
is made, calling the right function depending on the Python version.

It will likely not work with Python 3, but however applets are not
supposed to in the first place.

CCMAIL: release-team@kde.org
CCMAIL: kde-packager@kde.org

(cherry picked from commit 14f7b6a18ef8a81237df43551ee7bda122827d93)