Fix Removable property in soliddevice engine (HAL)
Current hack to check for Removable property in StorageAccess devices
goes up only one level to search for the StorageDrive device. This
works fine with UDev, but not with HAL, which can have (at least on
FreeBSD) intermediate devices between the volume and its StorageDrive
ancestor, for example partition-type devices. A generic solution would
be to create a method in Solid::Device, but this cannot be done until
libsolid2 is ready; meanwhile, let's add a getAncestorAs() function to
the engine.
The patch was submitted by Andriy Gapon and modified by me to handle
other similar cases.

While here, add logic to really set Hotpluggable property.

The final version of the patch was reviewed by Aaron Seigo on IRC.

REVIEW: 105432
REVIEW: 106378
CCMAIL: avg@FreeBSD.org