Fix: crash of plasma during resizing of panel
There is a strange behaviour when width of arrow area may be set to zero.
If user resizes panel then plasma will crash (somewhere in QtDeclarative
and QtScript). There is more strange thing: before crashing of plasma
the systrem tray can handle items from other applets.
So, there may be better way to fix (or this is a problem of Qt) but
this workaround (this commit) really fixes problem I think.