Fix PowerDevil generateProfiles to properly detect if toRam is available
The problem is that the DBus interface of powerdevil is being setup AFTER
generateProfile is executed, so libsolid will fail to query powerdevil
for the available suspend methods because powerdevil haven't made them
public yet.

The approach in the fix:
There are many ways of fixing this, the simpler one is to execute the dbus registering before generateProfiles is executed,
the others are based on the same principle of moving the current code around to make libsolid work.

REVIEW: 106859
(cherry picked from commit 936bd7f9c8ca9d49c19b4f8ab9e4b67cd4e9d72c)