new toolbox dismissing logic, toolbox follows button
Using a MouseEventListener causes problems passing the pressed event to
the underlying QML views, as our mouseventlistener accepts events but
passed them on to its children, so they still don't arrive at siblings.
So we use a MouseArea, blocking clicks from the Containment while the
toolbox is open, and making the mousearea invisible when the toolbox is
closed.

This commit also contains the logic for placing the toolbox close to the
button. A nice example why I love QML, it can be done cleanly in one
place, with a little bit of very easy javascript. Yay. :)