A modal approach to containment setup
Lock/Unlock widgets switches between "editing mode" and "content mode".
Controls and the applet fade in and out depending on the mode.
Implementation is a bit rough, but it mostly works.

The input model needs a bit of work, do we use the main content area for
overlays? Only slightly fade out the applet to emphasize the controls
more?